pyproject.toml
[project]
name = "devhub-cli"
version = "0.1"
description = "CLI interface to devhub"
readme = "README.md"
authors = [{name = "Daniel Rust"}]
license = {text = "Apache-2.0"}
requires-python = ">=3.8"
classifiers = []
dependencies = [
    "click ~=8.1.3",
    "python-dotenv ~=1.0.0",
    "djlint ~=1.34.0",
    "pyyaml ~=6.0",
    "requests-oauthlib ~=2.0.0",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project.urls]
Homepage = "https://github.com/devhub/devhub-cli"
Changelog = "https://github.com/devhub/devhub-cli/releases"
Issues = "https://github.com/devhub/devhub-cli/issues"
CI = "https://github.com/devhub/devhub-cli/actions"

[project.scripts]
devhub = "devhub.cli:cli"

[project.optional-dependencies]
test = ["pytest"]

[tool.setuptools.packages.find]
where = ["."]
include = ["devhub*"]
